refactor(ir): widen FixedValue Nint/Uint to i128/u64 for host-independent literals
FixedValue::Nint(isize)/Uint(usize) made IR literals platform-sized. Widen to
Nint(i128)/Uint(u64) so generation is host-independent, and encode nints through
the width-correct write_negative_integer_sz(i128, Sz::canonical(magnitude)) path,
which — unlike write_negative_integer(i64) — can encode i64::MIN (cbor_event#9).
Sz::canonical(magnitude) reproduces the default endpoint's canonical bytes exactly,
so output stays byte-identical for every value the old path encoded correctly.

Construction sites in parsing.rs cast at the cddl-AST boundary (upstream stays
isize/usize); the serialize.rs nint guard drops the now-moot isize::BITS clause;
generation/mod.rs casts the Tagged tag literal. GroupParsingType gains the
established #[allow(clippy::large_enum_variant)] (i128's 16-byte alignment grew
RustType, tipping the variant-size-gap lint on a short-lived, never-bulk-stored enum).

Tests: a to_bytes byte-identity unit test pins canonical nint bytes across the
magnitude ladder plus i64::MIN (expected bytes hard-coded, not re-derived from the
old path); a generation test pins the emitted i64::MIN serialize line under both
default and --preserve-encodings.

+++ src/generation/mod.rs
@@ -1933,7 +1933,7 @@ fn encoding_fields_impl(
             let mut encs = encoding_fields_impl(
                 types,
                 &format!("{name}_tag"),
-                (&ConceptualRustType::Fixed(FixedValue::Uint(*tag))).into(),
+                (&ConceptualRustType::Fixed(FixedValue::Uint(*tag as u64))).into(),
                 cli,
             );
             encs.append(&mut encoding_fields_impl(types, name, *child, cli));
